+++ blue-core/src/main/java/blue/scripting/JavaScriptProxyListener.java
@@ -1,0 +1,5 @@
+package blue.scripting;
+
+public interface JavaScriptProxyListener {
+    void javascriptProxyReinitializePerformed();
+}